backend_django/apps/accounts/permissions.py
from rest_framework.permissions import BasePermission


class IsAdminRole(BasePermission):
    def has_permission(self, request, _view):
        return bool(request.user and request.user.is_authenticated and request.user.role == "admin")
