backend_django/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.5.0
channels==4.2.0
pyotp==2.9.0
qrcode==8.0
Pillow==11.1.0
python-dotenv==1.0.1
